Update printer-fabtotum.config adjust servo pwm according to servo.attach() standard settings
To gain the same behavior for the probe as with fablin the angles have been corrected with the last commit. This commit adjusts PWM times to the default values of the Servo.attach() method to get the same behavior.
In the future one could figure the type of servo and adjust the angles and PWM settings to be more accurate. (180° opening with correct PWM)